Pull `lib` code that requires the app into `app/lib/audit`
Inside `app` everything is auto/eager-loaded by Rails, so you never need
to require anything else that lives inside `app`.

- Reduce unnecessary requires
- Clarify delegation to `PreservationCatalog` is not inside `Audit` module

This means less `require` statements overall.
