+++ app/lib/audit.rb
@@ -1,0 +1,10 @@
+require 'active_record_utils'
+require 'druid-tools'
+require 'profiler'
+
+require_relative 'audit/catalog_to_moab'
+require_relative 'audit/checksum'
+require_relative 'audit/moab_to_catalog'
+
+# @example Usage
+#  Audit::Checksum.validate_disk(storage_root)
