+++ spec/lib/audit/catalog_to_moab_class_spec.rb
@@ -1,8 +1,7 @@
 require 'rails_helper'
-require_relative '../../../lib/audit/catalog_to_moab.rb'
 require_relative '../../load_fixtures_helper.rb'
 
-RSpec.describe CatalogToMoab do
+RSpec.describe Audit::CatalogToMoab do
   let(:last_checked_version_b4_date) { (Time.now.utc - 1.day).iso8601 }
   let(:storage_dir) { 'spec/fixtures/storage_root01/moab_storage_trunk' }
   let(:limit) { Settings.c2m_sql_limit }
